src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
asgiref==3.9.1
    # via opentelemetry-instrumentation-asgi
azure-core==1.35.0
    # via
    #   azure-core-tracing-opentelemetry
    #   azure-identity
    #   azure-monitor-opentelemetry
    #   azure-monitor-opentelemetry-exporter
    #   msrest
azure-core-tracing-opentelemetry==1.0.0b12
    # via azure-monitor-opentelemetry
azure-identity==1.23.1
    # via
    #   aca-chaos-lab (pyproject.toml)
    #   azure-monitor-opentelemetry-exporter
azure-monitor-opentelemetry==1.6.12
    # via aca-chaos-lab (pyproject.toml)
azure-monitor-opentelemetry-exporter==1.0.0b40
    # via azure-monitor-opentelemetry
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
cryptography==45.0.5
    # via
    #   azure-identity
    #   msal
    #   pyjwt
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
fastapi==0.116.1
    # via aca-chaos-lab (pyproject.toml)
fixedint==0.1.6
    # via azure-monitor-opentelemetry-exporter
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via aca-chaos-lab (pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.6.1
    # via opentelemetry-api
isodate==0.7.2
    # via msrest
msal==1.33.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
msrest==0.7.1
    # via azure-monitor-opentelemetry-exporter
oauthlib==3.3.1
    # via requests-oauthlib
opentelemetry-api==1.31.1
    # via
    #   azure-core-tracing-opentelemetry
    #   azure-monitor-opentelemetry-exporter
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-psycopg2
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.52b1
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-psycopg2
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-asgi==0.52b1
    # via opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-dbapi==0.52b1
    # via opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-django==0.52b1
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-fastapi==0.52b1
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-flask==0.52b1
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-psycopg2==0.52b1
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-requests==0.52b1
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-urllib==0.52b1
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-urllib3==0.52b1
    # via azure-monitor-opentelemetry
opentelemetry-instrumentation-wsgi==0.52b1
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-flask
opentelemetry-resource-detector-azure==0.1.5
    # via azure-monitor-opentelemetry
opentelemetry-sdk==1.31.1
    # via
    #   azure-monitor-opentelemetry
    #   azure-monitor-opentelemetry-exporter
    #   opentelemetry-resource-detector-azure
opentelemetry-semantic-conventions==0.52b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.52b1
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-flask
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-instrumentation-urllib
    #   opentelemetry-instrumentation-urllib3
    #   opentelemetry-instrumentation-wsgi
packaging==25.0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-flask
psutil==7.0.0
    # via azure-monitor-opentelemetry-exporter
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   aca-chaos-lab (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via aca-chaos-lab (pyproject.toml)
pyjwt==2.10.1
    # via msal
python-dotenv==1.1.1
    # via pydantic-settings
redis==6.2.0
    # via aca-chaos-lab (pyproject.toml)
requests==2.32.4
    # via
    #   azure-core
    #   msal
    #   msrest
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via msrest
six==1.17.0
    # via azure-core
sniffio==1.3.1
    # via anyio
starlette==0.47.2
    # via fastapi
typing-extensions==4.14.1
    # via
    #   azure-core
    #   azure-identity
    #   fastapi
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via aca-chaos-lab (pyproject.toml)
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-dbapi
    #   opentelemetry-instrumentation-urllib3
zipp==3.23.0
    # via importlib-metadata
